reportのmigrationファイルの変更

+++ db/migrate/20210119122442_create_reports.rb
@@ -3,7 +3,7 @@ class CreateReports < ActiveRecord::Migration[5.2]
     create_table :reports do |t|
       t.string :title, null: false
       t.text :content
-      t.integer :time, null: false
+      t.float :time, null: false
       t.date :date, null: false
       t.string :place
       t.text :image
